Extract rotation host screen building blocks
